changed type of average rating to float. added rating to borrow record cause after each borrow, you should be able to review the book. Finished database Validation for book

+++ Database/database_objects/releases/v0.1.0/v0.1.0.sql
@@ -189,4 +189,4 @@ ADD COLUMN `status` VARCHAR(50) NOT NULL AFTER `return_date`;
 
 --changeset arman:ar3
 ALTER TABLE `scrum_library`.`borrow_record`
-ADD COLUMN `rating` INT AFTER `status`;
+ADD COLUMN `rating` DOUBLE  AFTER `status`;
